feat: create client app with Client and Supplier models
This commit introduces a new Django app called 'client' to manage
client and supplier information. It includes models, migrations,
admin configuration, and proxy models for workflow integration.

The models include fields for Xero integration, contact details,
and additional metadata. The Supplier model is a proxy model
extending the Client model.

The workflow app's Client and Supplier models are now proxy models
that inherit from the client app's models. This change centralizes
the Client and Supplier models in the 'client' app and allows other
apps to extend them.
